# swagger/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Api Rest Crud Prueba
  description: This is a simple API
  contact:
    email: ilkerfrancisco@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/ILKERFRANCISCO/Api_db/1.0.0
  description: SwaggerHub API Auto Mocking
tags:
- name: usuarios
  description: endpoints de usuarios
- name: publicaciones
  description: Operations available to regular developers
paths:
  /inventory:
    get:
      tags:
      - publicaciones
      summary: searches inventory
      description: |
        By passing in the appropriate options, you can search for
        available inventory in the system
      operationId: search_inventory
      parameters:
      - name: searchString
        in: query
        description: pass an optional search string for looking up inventory
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: skip
        in: query
        description: number of records to skip for pagination
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
          format: int32
      - name: limit
        in: query
        description: maximum number of records to return
        required: false
        style: form
        explode: true
        schema:
          maximum: 50
          minimum: 0
          type: integer
          format: int32
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryItem'
                x-content-type: application/json
        "400":
          description: bad input parameter
      x-openapi-router-controller: swagger_server.controllers.publicaciones_controller
    post:
      tags:
      - usuarios
      summary: adds an inventory item
      description: Adds an item to the system
      operationId: add_inventory
      requestBody:
        description: Inventory item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItem'
      responses:
        "201":
          description: succesfully
        "400":
          description: invalid user
        "409":
          description: the user doesnt exits
      x-openapi-router-controller: swagger_server.controllers.usuarios_controller
  /login:
    post:
      tags:
      - usuarios
      summary: adds an inventory item
      description: logs a user
      operationId: loginuser
      requestBody:
        description: Inventory item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItem'
      responses:
        "201":
          description: succesfully
        "400":
          description: invalid user
        "409":
          description: the user doesnt exits
      x-openapi-router-controller: swagger_server.controllers.usuarios_controller
  /logout:
    post:
      tags:
      - usuarios
      summary: adds an inventory item
      description: logs out a user
      operationId: logoutuser
      requestBody:
        description: Inventory item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItem'
      responses:
        "201":
          description: succesfully
        "400":
          description: invalid user
        "409":
          description: the user doesnt exits
      x-openapi-router-controller: swagger_server.controllers.usuarios_controller
  /register:
    post:
      tags:
      - usuarios
      summary: adds an inventory item
      description: register a user
      operationId: register_user
      requestBody:
        description: Inventory item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItem'
      responses:
        "201":
          description: succesfully
        "400":
          description: invalid user
        "409":
          description: the user doesnt exits
      x-openapi-router-controller: swagger_server.controllers.usuarios_controller
  /protected:
    post:
      tags:
      - usuarios
      summary: adds an inventory item
      description: protect a user
      operationId: protect_user
      requestBody:
        description: Inventory item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItem'
      responses:
        "201":
          description: succesfully
        "400":
          description: invalid user
        "409":
          description: the user doesnt exits
      x-openapi-router-controller: swagger_server.controllers.usuarios_controller
  /crear:
    post:
      tags:
      - usuarios
      summary: create an user
      description: create an user
      operationId: create_user
      requestBody:
        description: Inventory item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItem'
      responses:
        "201":
          description: succesfully
        "400":
          description: invalid user
        "409":
          description: the user doesnt exits
      x-openapi-router-controller: swagger_server.controllers.usuarios_controller
  /consultar:
    get:
      summary: Obtiene una lista
      operationId: consultar_get
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.default_controller
    post:
      tags:
      - usuarios
      summary: adds an inventory item
      description: consult an user
      operationId: query_user
      requestBody:
        description: Inventory item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItem'
      responses:
        "201":
          description: succesfully
        "400":
          description: invalid user
        "409":
          description: the user doesnt exits
      x-openapi-router-controller: swagger_server.controllers.usuarios_controller
  /eliminar:
    post:
      tags:
      - usuarios
      summary: adds an inventory item
      description: delete an user
      operationId: delete_user
      requestBody:
        description: Inventory item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItem'
      responses:
        "201":
          description: succesfully
        "400":
          description: invalid user
        "409":
          description: the user doesnt exits
      x-openapi-router-controller: swagger_server.controllers.usuarios_controller
  /update:
    post:
      tags:
      - usuarios
      summary: adds an inventory item
      description: update an user
      operationId: update_user
      requestBody:
        description: Inventory item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItem'
      responses:
        "201":
          description: succesfully
        "400":
          description: invalid user
        "409":
          description: the user doesnt exits
      x-openapi-router-controller: swagger_server.controllers.usuarios_controller
components:
  schemas:
    InventoryItem:
      required:
      - id
      - manufacturer
      - name
      - releaseDate
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        name:
          type: string
          example: Widget Adapter
        releaseDate:
          type: string
          format: date-time
          example: 2016-08-29T09:12:33.001Z
        manufacturer:
          $ref: '#/components/schemas/Manufacturer'
      example:
        releaseDate: 2016-08-29T09:12:33.001Z
        name: Widget Adapter
        id: d290f1ee-6c54-4b01-90e6-d701748f0851
        manufacturer:
          phone: 408-867-5309
          name: ACME Corporation
          homePage: https://www.acme-corp.com
    Manufacturer:
      required:
      - name
      type: object
      properties:
        name:
          type: string
          example: ACME Corporation
        homePage:
          type: string
          format: url
          example: https://www.acme-corp.com
        phone:
          type: string
          example: 408-867-5309
      example:
        phone: 408-867-5309
        name: ACME Corporation
        homePage: https://www.acme-corp.com